requirements.txt
# Bibliotecas utilizadas
pyautogui
selenium
pandas
python-dotenv